+++ packages/lumx-react/src/index.ts
@@ -38,7 +38,13 @@ export { InputHelper, InputHelperProps } from './components/input-helper/InputHe
 
 export { InputLabel, InputLabelProps } from './components/input-label/InputLabel';
 
-export { Thumbnail, ThumbnailProps, ThumbnailVariant, ThumbnailAspectRatio } from './components/thumbnail/Thumbnail';
+export {
+    Thumbnail,
+    ThumbnailProps,
+    ThumbnailVariant,
+    ThumbnailAspectRatio,
+    CrossOrigin,
+} from './components/thumbnail/Thumbnail';
 
 export { FocusPoint } from './components/thumbnail/FocusedImageOptions';